add CategoriaServices + alter. Id: Long => Integer

+++ src/main/java/com/ayvy/api_java/business/ClienteService.java
@@ -44,7 +44,7 @@ public class ClienteService {
     }
 
     //Atualizar os dados do cliente 'Update':
-    public void atualizarClientePorId(Long id, Cliente cliente){
+    public void atualizarClientePorId(Integer id, Cliente cliente){
         Cliente clienteEntity = repository.findById(id).orElseThrow(
                 () -> new RuntimeException("Usuário/cliente não encontrado")
         );
