+++ src/main/java/com/ayvy/api_java/infrastructure/entitys/Produto.java
@@ -21,11 +21,16 @@ public class Produto {
     //Cria um valor automáticamente:
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
 
     @Column(name = "nome_produto")
     private String nome_produto;
 
     @Column(name = "preco")
     private Integer preco;
+
+    //Faltam:
+    //Descrição String
+    //data_criação LocalDateTime
+    //preço deve ser BIGDECIMAL!!
 }
